perlin noise map generation seeded and generated on server/client

+++ server/include/packet/buffer/PacketWriter.hpp
@@ -12,7 +12,7 @@ class PacketWriter {
     size_t m_offset;
 
     void writeU8(uint8_t x);
-    void writeU16(uint8_t x);
+    void writeU16(uint16_t x);
     void writeU32(uint32_t x);
     void writeFloat(float x);
     void writeString(const std::string& x);
